requirements.txt
fastapi
uvicorn
python-multipart
jinja2
google-cloud-documentai
google-api-core
pdfplumber
openpyxl
pyjwt
cryptography
requests
pypdf
redis
google-cloud-firestore
firebase-admin
google-cloud-bigquery
